refactor: TripFacade 도입 — 멀티 도메인 조회 로직 분리
